Moving class to different files to be compliant with StyleCop analyzers
