+++ libraries/Microsoft.Bot.Builder.Ai.Translation/ILocaleConverter.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-
 namespace Microsoft.Bot.Builder.Ai.Translation
 {
     public interface ILocaleConverter
