+++ libraries/Microsoft.Bot.Builder.Ai.Translation/TextAndDateTime.cs
@@ -2,17 +2,9 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text.RegularExpressions;
-using Microsoft.Recognizers.Text;
-using Microsoft.Recognizers.Text.DateTime;
 
 namespace Microsoft.Bot.Builder.Ai.Translation
 {
-
     /// <summary>
     /// TextAndDateTime Class used to store  text and date time object
     /// from Microsoft Recognizer recognition result.
